parsed parts to appropritate go files
